Semana 06
Fix ahora funciona el Productos, aunque aún no se muestran el Listado de Productos (Queda el Arreglo Pendiente)

+++ src/main/java/org/jared/jaredctienda/services/FirebaseStorageService.java
@@ -14,13 +14,10 @@ import org.springframework.web.multipart.MultipartFile;
 
 @Service
 public class FirebaseStorageService {
-
     @Value("${firebase.bucket.name}")
     private String bucketName;
-
     @Value("${firebase.storage.path}")
     private String storagePath;
-
     // Aquí se manejaría la inyección del cliente de Storage como un bean
     private final Storage storage;
 
@@ -29,17 +26,18 @@ public class FirebaseStorageService {
     }
 
     //Sube un archivo de imagen al almacenamiento de Firebase.
-    public String uploadImage(MultipartFile localFile, String folder, Long id) throws IOException {
+    public String uploadImage(MultipartFile localFile, String folder, Integer id) throws IOException {
         String originalName = localFile.getOriginalFilename();
         String fileExtension = "";
-
         if (originalName != null && originalName.contains(".")) {
             fileExtension = originalName.substring(originalName.lastIndexOf("."));
         }
 
         // Se genera el nombre del archivo con un formato consistente.
         String fileName = "img" + getFormattedNumber(id) + fileExtension;
+
         File tempFile = convertToFile(localFile);
+
         try {
             return uploadToFirebase(tempFile, folder, fileName);
         } finally {
@@ -52,7 +50,6 @@ public class FirebaseStorageService {
 
     //Convierte un MultipartFile a un archivo temporal en el servidor.
     private File convertToFile(MultipartFile multipartFile) throws IOException {
-
         File tempFile = File.createTempFile("upload-", ".tmp");
         try (FileOutputStream fos = new FileOutputStream(tempFile)) {
             fos.write(multipartFile.getBytes());
@@ -62,7 +59,6 @@ public class FirebaseStorageService {
 
     //Sube el archivo al almacenamiento de Firebase y genera una URL firmada.
     private String uploadToFirebase(File file, String folder, String fileName) throws IOException {
-
         // Definimos el ID del blob y su información
         BlobId blobId = BlobId.of(bucketName, storagePath + "/" + folder + "/" + fileName);
         String mimeType = Files.probeContentType(file.toPath());
@@ -80,7 +76,6 @@ public class FirebaseStorageService {
      * Genera un string numérico con un formato de 14 dígitos, rellenado con
      * ceros a la izquierda.
      */
-
     private String getFormattedNumber(long id) {
         return String.format("%014d", id);
     }
